requirements.txt
asgiref==3.8.1
django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
whitenoise==6.7.0
gunicorn==21.2.0
setuptools
wheel
PyMySQL==1.1.1
openai==0.28.1
